selectors file has been created
