+++ src/Components/CommandPanel.js
@@ -7,7 +7,8 @@ import {
   showAll,
   filterCompleted,
 } from '../features/filterSlice';
-import { removeCompleted, selectTodosLeftCount } from '../features/todosSlice';
+import { removeCompleted } from '../features/todosSlice';
+import { selectTodosLeftCount } from '../features/selectTodos';
 
 export default function CommandPanel() {
   const todosLeftCount = useSelector(selectTodosLeftCount);
