+++ src/features/filterSlice.js
@@ -1,8 +1,6 @@
 import { createSlice } from '@reduxjs/toolkit';
 
-const FILTER_COMPLETED = 'FILTER_COMPLETED';
-const FILTER_ACTIVE = 'FILTER_ACTIVE';
-const SHOW_ALL = 'SHOW_ALL';
+import { SHOW_ALL, FILTER_COMPLETED, FILTER_ACTIVE } from '../globalConst';
 
 const filterSlice = createSlice({
   name: 'filter',
@@ -15,17 +13,5 @@ const filterSlice = createSlice({
 });
 
 export const { filterActive, showAll, filterCompleted } = filterSlice.actions;
-
-export const filteredTodos = state => {
-  const { filter, todos } = state;
-  switch (filter) {
-    case FILTER_ACTIVE:
-      return todos.filter(todo => !todo.done);
-    case FILTER_COMPLETED:
-      return todos.filter(todo => todo.done);
-    default:
-      return todos;
-  }
-};
 
 export default filterSlice.reducer;
