+++ src/features/localStorageMiddleware.js
@@ -1,11 +1,10 @@
-export const KEY_TODO = 'KEY_TODO';
+import { KEY_TODO } from '../globalConst';
 
 const localStorageMiddleware = store => next => action => {
-  const result = next(action);
   if (action.type.startsWith('todo')) {
     localStorage.setItem(KEY_TODO, JSON.stringify(store.getState().todos));
   }
-  return result;
+  return next(action);
 };
 
 export default localStorageMiddleware;
