+++ src/features/todosSlice.js
@@ -46,8 +46,4 @@ const todosSlice = createSlice({
 export const { addTodo, todoDone, todoRemove, removeCompleted } =
   todosSlice.actions;
 
-export const selectTodos = state => state.todos;
-export const selectTodosLeftCount = state =>
-  state.todos.filter(todo => !todo.done).length;
-
 export default todosSlice.reducer;
