+++ src/globalConst.js
@@ -1,0 +1,5 @@
+export const FILTER_COMPLETED = 'FILTER_COMPLETED';
+export const FILTER_ACTIVE = 'FILTER_ACTIVE';
+export const SHOW_ALL = 'SHOW_ALL';
+
+export const KEY_TODO = 'KEY_TODO';
